requirements.txt
openai>=1.0.0
python-dotenv>=1.0.0
dataclasses>=0.8; python_version < "3.7"
